[R40] Update tools repository to ROM40. Fix bug in recoverylink.py

+++ releasetools/recoverylink.py
@@ -12,17 +12,18 @@ try:
     for line in linelist:
         line = line.rstrip()
         filepath = line.split('|')
-        filepath[0] = filepath[0].replace('system', 'SYSTEM')
-        filepath[1] = filepath[1].replace('system', 'SYSTEM')
-        rm = 'rm -f ' + path+ '/' + filepath[0]
-        os.popen(rm)
-        dirname=os.path.dirname(filepath[0])
-        filepath[0] = os.path.basename(filepath[0])
-        filepath[1] = os.path.basename(filepath[1])
-        #ln = 'ln -s '+ path+ '/'+ filepath[1] + ' ' + path+ '/'+  filepath[0]
-        ln = 'cd ' + path + '/' + dirname + ';' + 'ln -s ' + filepath[1] + ' ' +  filepath[0]
-        #print ln
-        os.popen(ln)
+        if len(filepath) >= 2:
+            filepath[0] = filepath[0].replace('system', 'SYSTEM')
+            filepath[1] = filepath[1].replace('system', 'SYSTEM')
+            rm = 'rm -f ' + path+ '/' + filepath[0]
+            os.popen(rm)
+            dirname=os.path.dirname(filepath[0])
+            filepath[0] = os.path.basename(filepath[0])
+            filepath[1] = os.path.basename(filepath[1])
+            #ln = 'ln -s '+ path+ '/'+ filepath[1] + ' ' + path+ '/'+  filepath[0]
+            ln = 'cd ' + path + '/' + dirname + ';' + 'ln -s ' + filepath[1] + ' ' +  filepath[0]
+            #print ln
+            os.popen(ln)
 except IOError:
     print r"%s isn't exist" % linkfile_path
     sys.exit(1)
